requirements.txt
aiogram>=3.0.0
google-generativeai>=0.3.0
mistralai>=1.0.0
chromadb>=0.3.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
networkx>=3.0
numpy>=1.24.0
